# WebAPI/Models/Registeration.cs
using System;
namespace WebAPI.Models
{
    public class Registration
    {
        public int Id { get; set; }
        public string? Name { get; set; }
        public string? Email { get; set; }
        public string? Password { get; set; }
    }
}